Refactor: Photocard 컴포넌트의 공통적인 부분을 PhotoCardTemplate으로 추상화
