+++ client/src/component/list/photo/PhotoListCard.scss
@@ -46,6 +46,7 @@
   // gap: 3em;
 
   .PhotoCard { margin: 1.5em; }
+  .VoucherCard { margin: 1.5em; }
 }
 
 @media screen and (max-width: 25rem) {
