+++ client/src/component/list/photo/PhotoListCard.tsx
@@ -7,12 +7,11 @@ import SearchInput from './content/SearchInput';
 import SearchLabelList from './content/SearchLabelList';
 import FilterCheck from './content/FilterCheck';
 import PhotoList from './content/PhotoList';
-import { PhotoType } from '@api/photoAPI';
 import './PhotoListCard.scss';
 
 interface PhotoListCardProps {
   icon?: IconDefinition;
-  handleClickIcon?: (photo: PhotoType) => void;
+  handleClickIcon?: (photocardId: number) => void;
   children?: React.ReactNode;
 }
 const PhotoListCardDefaultProps = {};
