+++ client/src/component/list/voucher/VoucherListCard.tsx
@@ -1,4 +1,5 @@
 import React, { useEffect } from 'react';
+import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { useAppDispatch } from '@app/redux/reduxHooks';
 import Card from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
@@ -11,11 +12,13 @@ import { initialize } from './voucherListSlice';
 import '../photo/PhotoListCard.scss';
 
 interface VoucherListCardProps {
+  icon?: IconDefinition;
+  handleClickIcon?: (voucherId: number) => void;
   children?: React.ReactNode;
 }
 const VoucherListCardDefaultProps = {};
 
-function VoucherListCard({ children }: VoucherListCardProps & typeof VoucherListCardDefaultProps) {
+function VoucherListCard({ icon, handleClickIcon, children }: VoucherListCardProps & typeof VoucherListCardDefaultProps) {
   const dispatch = useAppDispatch();
 
   // 언마운트시 리덕스 상태 초기화
@@ -33,7 +36,7 @@ function VoucherListCard({ children }: VoucherListCardProps & typeof VoucherList
       </CardHeader>
       <CardBody>
         <FilterCheck />
-        <VoucherList />
+        <VoucherList icon={icon} handleClickIcon={handleClickIcon} />
       </CardBody>
     </Card>
   );
