+++ client/src/page/admin/voucher/list/Index.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useCallback } from 'react';
+import { faArrowRight } from '@fortawesome/free-solid-svg-icons';
 import VoucherListCard from '@component/list/voucher/VoucherListCard';
 import './Index.scss';
 
@@ -8,10 +9,14 @@ interface IndexProps {
 const IndexDefaultProps = {};
 
 function Index({ children }: IndexProps & typeof IndexDefaultProps) {
+  const handleClickDetailicon = useCallback((voucherId: number) => {
+    console.log(`TODO: ${voucherId} 소유권 상세 페이지로 이동`);
+  }, []);
+
   return (
     <div className="VoucherListPage">
       <h1 className="title-label">소유권 목록</h1>
-      <VoucherListCard />
+      <VoucherListCard icon={faArrowRight} handleClickIcon={handleClickDetailicon} />
     </div>
   );
 }
